ReadをModel化する
